requirements.txt
pymongo==4.15.4
pandas==2.3.3
python-docx==1.2.0
numpy==2.3.5
